fix: Rust compilation errors + add keybind instructions to README
Fixed all remaining compilation errors in Rust version:

**VadMode Type Conversion (E0277):**
- webrtc_vad 0.4 uses VadMode enum, not i32 directly
- Added proper match statement to convert i32 -> VadMode
- VadMode::Quality (0), LowBitrate (1), Aggressive (2), VeryAggressive (3)

**VAD Error Display (E0277):**
- is_voice_segment() returns Err(()) not a Display-able error
- Changed debug!("VAD error: {}", e) to debug!("VAD error: invalid frame length")
- Now handles () error type correctly

**Type Mismatch (E0308):**
- Fixed remaining vad_aggressiveness u8->i32 cast in test command (main.rs:243)
- Now all AudioRecorder::new calls use `.as i32` cast

**Removed Unused Imports:**
- audio/mod.rs: Removed SampleFormat, Stream
- streaming/mod.rs: Removed Context, HostTrait
- transcribe/mod.rs: Removed warn

**Documentation:**
- Added keybinding setup instructions to rust-scribe/README.md
- Includes KDE Plasma, GNOME, and other DE instructions
- Uses scribe-rust command with proper PATH handling

All compiler errors resolved. Code now builds successfully.

+++ rust-scribe/src/main.rs
@@ -240,7 +240,7 @@ fn test_system(config: &Config) -> Result<()> {
     let mut audio_recorder = AudioRecorder::new(
         config.audio.sample_rate,
         config.audio.channels as u16,
-        config.audio.vad_aggressiveness,
+        config.audio.vad_aggressiveness as i32,
         config.audio.silence_duration,
         config.audio.min_audio_duration,
     )?;
